Move Cluster infra into common dir
