+++ openstack/cce/v1/nodes/results.go
@@ -126,8 +126,6 @@ func (r GetResult) Extract() (*GetNode, error) {
 	return s, err
 }
 
-// UpdateResult represents the result of an update operation. Call its Extract
-// method to interpret it as a Node.
 type UpdateResult struct {
 	golangsdk.Result
 }
